requirements.txt
torch>=2.1.0
transformers>=4.36.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.0
loguru>=0.7.0
manim>=0.18.0
pandas>=2.1.0